Split player file into player and strategies files
