+++ axelrod/ultimatum/__init__.py
@@ -1,2 +1,3 @@
 
 from .player import *
+from .strategies import *
